requirements.txt
ansible-core>=2.17,<3.0
python-telegram-bot[webhooks]>=21.0,<23.0

